+++ docs/src/examples/components/Avatar/Types/AvatarExample.shorthand.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import { Avatar } from 'stardust'
+
+const AvatarExampleShorthand = () => (
+  <div style={{ background: 'white' }}>
+    <Avatar name="John Doe" />
+  </div>
+)
+
+export default AvatarExampleShorthand