+++ docs/src/examples/components/Avatar/Types/index.tsx
@@ -1,0 +1,15 @@
+import React from 'react'
+import ComponentExample from 'docs/src/components/ComponentDoc/ComponentExample'
+import ExampleSection from 'docs/src/components/ComponentDoc/ExampleSection'
+
+const Types = () => (
+  <ExampleSection title="Types">
+    <ComponentExample
+      title="Default"
+      description="A default Avatar."
+      examplePath="components/Avatar/Types/AvatarExample"
+    />
+  </ExampleSection>
+)
+
+export default Types