+++ docs/src/examples/components/Avatar/Variations/AvatarExampleName.shorthand.tsx
@@ -1,0 +1,6 @@
+import React from 'react'
+import { Avatar } from 'stardust'
+
+const AvatarExampleNameShorthand = () => <Avatar status="Available" name="John Doe" />
+
+export default AvatarExampleNameShorthand