+++ docs/src/examples/components/Avatar/Variations/AvatarExampleSrc.shorthand.tsx
@@ -1,0 +1,6 @@
+import React from 'react'
+import { Avatar } from 'stardust'
+
+const AvatarExampleSrcShorthand = () => <Avatar src="/public/images/avatar/small/matt.jpg" />
+
+export default AvatarExampleSrcShorthand