+++ src/components/Avatar/avatarVariables.ts
@@ -1,0 +1,5 @@
+export default (siteVars: any) => {
+  return {
+    presenceIconPadding: 2,
+  }
+}