+++ src/components/Icon/fontAwesomeIconRules.ts
@@ -23,6 +23,9 @@ const fontAwesomeIcons = new Map([
   ['compass outline', 'f14e'],
   ['calendar alternate outline', 'f073'],
   ['thin circle', 'f111'],
+  ['check', 'f00c'],
+  ['minus', 'f068'],
+  ['clock', 'f017'],
 ])
 
 export default name => {