+++ src/components/Label/labelRules.ts
@@ -2,7 +2,7 @@ import { pxToRem } from '../../lib'
 
 export default {
   root: ({ props, variables }) => ({
-    padding: `${pxToRem(6)} ${pxToRem(8)}`,
+    padding: variables.padding,
     fontWeight: '500',
     backgroundColor: 'rgb(232, 232, 232)',
     color: 'rgba(0, 0, 0, 0.6)',