+++ test/specs/components/Avatar/Avatar-test.tsx
@@ -1,0 +1,7 @@
+import { isConformant } from 'test/specs/commonTests'
+
+import Avatar from 'src/components/Avatar/Avatar'
+
+describe('Avatar', () => {
+  isConformant(Avatar)
+})